style(entity-types): Rename files for resolving entities
use the name resolve instead of reducer, to differentiate from the reducer type
