+++ packages/data-point/lib/entity-types/entity-collection/resolve.js
