+++ packages/data-point/lib/entity-types/entity-entry/index.js
@@ -1,9 +1,9 @@
 /* eslint global-require: 0 */
 
 const factory = require('./factory')
-const reducer = require('./reducer')
+const resolve = require('./resolve').resolve
 
 module.exports = {
   create: factory.create,
-  resolve: reducer.resolve
+  resolve: resolve
 }
